+++ cr/patch/symfony/polyfill/src/Intl/Icu/IntlDateFormatter.php
@@ -221,7 +221,7 @@ abstract class IntlDateFormatter
      *
      * @param int|string|\DateTimeInterface $datetime The timestamp to format
      *
-     * @return string|bool The formatted value or false if formatting failed
+     * @return string|false The formatted value or false if formatting failed
      *
      * @see https://php.net/intldateformatter.format
      *
@@ -294,7 +294,7 @@ abstract class IntlDateFormatter
      *
      * @throws MethodNotImplementedException
      */
-    public function formatObject($datetime, $format = null, /*string */$locale = null)
+    public static function formatObject($datetime, $format = null, /*string */$locale = null)
     {
         $locale = backport_type_check('?string', $locale);